Separate statement logic into partial class
