+++ source/Databricks/source/SqlStatementExecution/Client/DatabricksSqlStatementClient.cs
@@ -39,7 +39,7 @@ namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution.Client;
 /// https://learn.microsoft.com/en-us/aspnet/core/fundamentals/http-requests?view=aspnetcore-7.0
 /// https://learn.microsoft.com/en-gb/azure/databricks/sql/api/sql-execution-tutorial
 /// </summary>
-public class DatabricksSqlStatementClient : IDatabricksSqlStatementClient
+public partial class DatabricksSqlStatementClient : IDatabricksSqlStatementClient
 {
     private const string StatementsEndpointPath = "/api/2.0/sql/statements";
     private readonly HttpClient _httpClient;
@@ -123,42 +123,6 @@ public class DatabricksSqlStatementClient : IDatabricksSqlStatementClient
         }
 
         _logger.LogDebug("SQL statement executed. Rows returned: {RowCount}", rowCount);
-    }
-
-    public async IAsyncEnumerable<dynamic> ExecuteStatementAsync(Abstractions.Statement statement, Format format)
-    {
-        var strategy = format.GetStrategy(_options);
-        var request = strategy.GetStatementRequest(statement);
-        var sw = Stopwatch.StartNew();
-        var response = await request.WaitForSqlWarehouseResultAsync(_httpClient, StatementsEndpointPath);
-        // Metrics.RecordWarehouseDuration(sw.Elapsed);
-
-        if (_httpClient.BaseAddress == null) throw new InvalidOperationException();
-
-        if (response.manifest.total_row_count <= 0)
-        {
-            yield break;
-        }
-
-        foreach (var chunk in response.manifest.chunks)
-        {
-            sw.Restart();
-            var uri = StatementsEndpointPath +
-                      $"/{response.statement_id}/result/chunks/{chunk.chunk_index}?row_offset={chunk.row_offset}";
-            var chunkResponse = await _httpClient.GetFromJsonAsync<ManifestChunk>(uri);
-            // Metrics.RecordChunkDuration(sw.Elapsed);
-
-            if (chunkResponse?.external_links == null) continue;
-
-            sw.Restart();
-            await using var stream = await _externalHttpClient.GetStreamAsync(chunkResponse.external_links[0].external_link);
-            // Metrics.RecordDurationOfDataRetrieval(sw.Elapsed);
-
-            await foreach (var row in strategy.ExecuteAsync(stream, response))
-            {
-                yield return row;
-            }
-        }
     }
 
     private async Task<SqlResponse> GetFirstChunkOrNullAsync(
